finished updating frontend

+++ backend/apps/events/helpers.py
@@ -3,8 +3,9 @@ from .models import Attendable, SlotDistribution
 
 def create_attendable(attendable_data, event):
     if (
-        attendable_data.has_key("price")
-        and attendable_data.has_key("binding_signup")
+        hasattr(attendable_data, "price")
+        and attendable_data.price != None
+        and hasattr(attendable_data, "binding_signup")
         and attendable_data.binding_signup == False
     ):
         raise ValueError("Betalt påmelding krever bindende påmelding")
@@ -22,15 +23,16 @@ def create_slot_distributions(slot_distribution_data, attendable):
 
     # Create parent distribution
     grade_dist_data = (
-        slot_distribution_data.pop("grade_years") if slot_distribution_data.has_key("grade_years") else None
+        slot_distribution_data.pop("grade_years") if hasattr(slot_distribution_data, "grade_years") else None
     )
+
     slot_distribution = SlotDistribution()
-    for k, v in slot_distribution.items():
+    for k, v in slot_distribution_data.items():
         setattr(slot_distribution, k, v)
     setattr(slot_distribution, "attendable", attendable)
 
     # Create child distributions if different grades have a specififc number of slots
-    if grade_dist_data is not None:
+    if grade_dist_data is not None and len(grade_dist_data) > 0:
         total_grades_allowed = ""
         total_child_slots = 0
         for dist in grade_dist_data:
@@ -47,32 +49,34 @@ def create_slot_distributions(slot_distribution_data, attendable):
         if total_child_slots != slot_distribution.available_slots:
             raise ValueError("Totalt antall plasser stemmer ikke overens med fordelingen av plasser")
         setattr(slot_distribution, "grade_years", total_grades_allowed)
+        setattr(slot_distribution, "available_slots", total_child_slots)
 
     slot_distribution.save()
     for child in child_dists:
         setattr(child, "parent_distribution", slot_distribution)
+        child.save()
 
     return slot_distribution
 
 
 def update_slot_distributions(slot_distribution_data, slot_distribution, has_grade_distributions):
     grade_dist_data = (
-        slot_distribution_data.pop("grade_years") if slot_distribution_data.has_key("grade_years") else None
+        slot_distribution_data.pop("grade_years") if hasattr(slot_distribution_data, "grade_years") else None
     )
     # Update parent distribution
-    for k, v in slot_distribution.items():
+    for k, v in slot_distribution_data.items():
         setattr(slot_distribution, k, v)
     slot_distribution.save()
 
     # Update child distributions
-    if grade_dist_data is not None:
-        current_no_of_categories = len(list(slot_distribution.child_distributions))
+    if grade_dist_data is not None and len(grade_dist_data) > 0:
+        children = list(slot_distribution.child_distributions.all())
+        current_no_of_categories = len(children)
         no_of_new_categories = len(grade_dist_data)
-        children = list(slot_distribution.child_distributions)
 
-        # Number of distributions is the same or fewer, must remove some children if fewer and then update their values
+        # Number of distributions is the same or fewer, must remove children if fewer and then update their values
         if current_no_of_categories >= no_of_new_categories:
-            for _ in range(no_of_new_categories - current_no_of_categories):
+            for _ in range(current_no_of_categories - no_of_new_categories):
                 child = children.pop()
                 child.delete()
 
@@ -86,7 +90,7 @@ def update_slot_distributions(slot_distribution_data, slot_distribution, has_gra
                 setattr(child, "grade_years", dist.category)
                 child.save()
 
-        else:  # Number of distributions has increased, must add some children
+        else:  # Number of distributions has increased, must add  children
             total_grades_allowed = ""
             total_child_slots = 0
             for dist in grade_dist_data[:current_no_of_categories]:
@@ -106,21 +110,21 @@ def update_slot_distributions(slot_distribution_data, slot_distribution, has_gra
                     grade_years=dist.category,
                     parent_distribution=slot_distribution,
                 )
+                child.save()
 
         total_grades_allowed = total_grades_allowed[:-1]
         listed_total_grades_allowed = [int(val) for val in total_grades_allowed.split(",")]
         if len(set(listed_total_grades_allowed)) != len(listed_total_grades_allowed):
             raise ValueError("Samme trinn kan ikke være i flere 'Antall plasser'-kategorier")
 
-        if total_child_slots != slot_distribution.available_slots:
-            raise ValueError("Totalt antall plasser stemmer ikke overens med fordelingen av plasser")
-
         setattr(slot_distribution, "grade_years", total_grades_allowed)
+        setattr(slot_distribution, "available_slots", total_child_slots)
+        slot_distribution.save()
 
     if (
-        len(slot_distribution.child_distributions) > 0 and not has_grade_distributions
+        len(list(slot_distribution.child_distributions.all())) > 0 and not has_grade_distributions
     ):  # No longer using categorical slot distributions, must delete all child distributions
-        for child in slot_distribution.child_distributions:
+        for child in slot_distribution.child_distributions.all():
             child.delete()
 
     return slot_distribution
